refactor(mall): 将订单状态统一定义到 protobuf
  - 新增 OrderStatus 订单状态枚举
  - 将订单相关状态字段改为枚举类型
  - 删除 app 层重复的订单状态常量
  - 适配数据库模型与 protobuf 状态类型
  - 更新 app、admin 和 mall 服务的类型转换

+++ cmd/admin/internal/logic/mall_admin/vars.go
@@ -62,7 +62,7 @@ func orderToType(o *pb.Order) types.AdminOrderResp {
 		OriginalAmount: o.OriginalAmount,
 		DiscountAmount: o.DiscountAmount,
 		PayAmount:      o.PayAmount,
-		Status:         o.Status,
+		Status:         int32(o.Status),
 		PayType:        o.PayType,
 		PayTime:        o.PayTime,
 		Remark:         o.Remark,
